Improve documentation around validation of opt config in discrete.py (#3331)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3331

The documentation in the utils and the general method was not super clear to me and we had to get insight from Sam to understand why this didn't fail if relativize transform was applied.

To me the oriignal documentaion reads as if no relative constraints may exist, but if relativize has been applied now all constraints are relative so really no constraints are relative. So it's really a mix of relative and absolute constraints that create the error

Reviewed By: lena-kashtelyan

Differential Revision: D69223913

fbshipit-source-id: b187ac261da181818705c72e209fe4271b067cd2