Remove potential SQL deprecation message (#3346)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3346

We added this in D64781681 to get a temperature check on whether or not this feature was being used in OSS. Based on activity on https://github.com/facebook/Ax/issues/2975 the answer is yes.

This is good information for our team to have and will help inform our plans for Ax in the future.

Reviewed By: lena-kashtelyan

Differential Revision: D69471648

fbshipit-source-id: 6638b56cf295fd4f7d29e4ca1880e381d106b698