Remove logspew from percentile es (#3366)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3366

Playing whack-a-mode with some logs while developing tutorials. Logs are in an unacceptable state, producing thousands of log entries which take up more than 99% of the page real estate before this diff. This is because we're multiple logs per progression call to `should_stop_trial_early` regardless of whether anything happened. We do not need this much visibility on whether or not we decide to kill a trial.

In the tutorials we're working on (ex N6570044 we were seeing thousands of logs before this change, now we see a couple dozen).

Changes:
* No longer log if there havent been enough completed trials to make an ES decision
* No longer log if no trials have met minimum observed progressions
* No longer log the value of every trial before we make an ES decision (this happened twice)
* No longer log at the beginning of ever ES decision that we are about to make an ES decision
* No longer log when we consider stopping a trial and chose not to (only log if we actually decide it should stop)
* [RFC] No longer log if an error happened during interpolation and we decide to use uninterpolated values instead (its not clear to me how important this error is -- fwiw in the tutorial we run into this very frequently -- if its a real actionable issue we need to fix it)

Reviewed By: saitcakmak, lena-kashtelyan, sunnyshen321

Differential Revision: D69616469

fbshipit-source-id: ceea18a40aa277f82314345ccb82d0c611cbcdcd