Fix logging verbosity in BackendSimulator (#3368)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3368

The verbose logging option in `BackendSimulator` was broken by D65737627, which sets the level of all Ax loggers to the level in `SchedulerOptions` when the schedule is instantiated. In a benchmark run, the logging level of the `BackendSimulator` starts out one thing then eventually gets changed.

This diff changes the logic so that `verbose_logging` controls the level at which logs are logged rather then at which they are visible. Previously, all logs were INFO and were not shown if the level was raised to warning; now, logs can be at either INFO or DEBUG so the level at which they are shown can still be controlled by Scheduler. Under Ax's default INFO logging level, this maintains the behavior that logs will only be surfaced when verbose_logging is True.

Reviewed By: lena-kashtelyan

Differential Revision: D69623149

fbshipit-source-id: 3c2a3660472525399d297b99c203a6fd7b38a44b