Simplify observations_from_data/map_data (#3339)
Summary:

This diff cleans up a rarely used & untested part of `observations_from_data`. The code had logic for splitting the dataframe into two, and processing the dfs separately. However, this part of the code was not tested and it wasn't being executed in rare cases where the feature columns included NaN / NaT values. The `_observations_from_dataframe` helper is slightly modified to support NaN / NaT values, eliminating the need for splitting the dfs.

Reviewed By: lena-kashtelyan

Differential Revision: D69419519