Simplify observations_from_data/map_data (#3339)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3339

This diff cleans up a rarely used & untested part of `observations_from_data`. The code had logic for splitting the dataframe into two, and processing the dfs separately. However, this part of the code was not tested and it wasn't being executed in rare cases where the feature columns included NaN / NaT values. The `_observations_from_dataframe` helper is slightly modified to support NaN / NaT values, eliminating the need for splitting the dfs.

Reviewed By: lena-kashtelyan

Differential Revision: D69419519

fbshipit-source-id: c95c88d08fb5245c8b8dd6505569bab54e105347