Remove parameter constraints when expanding model space in Adapter (#3408)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3408

By default we expand range parameters in the model space so that the
training data include all modelable data, not just those that fall within the
box bounds.

A similar argument suggests that we should also include training data that
violate parameter constraints in the model. They are modelable and can be used
to improve predictions at points that satisfy parameter constraints.

We do this by simply dropping parameter constraints from the model space, if
the model space is specified to be expanded.

This will also fix https://github.com/facebook/Ax/issues/1568 . The issue there
is that generated points violated parameter constraints due to numerical
issues, and so then were left out of the training data and repeatedly
generated. With this change, those points would be included in the training
data and so likely not re-generated. Note, however, that the purpose of this
change is not just to fix this issue; it is a change that is right because it
better fits the concept of the modeling space.

Reviewed By: sdaulton, saitcakmak

Differential Revision: D69880066

fbshipit-source-id: c5bc5d042b567eceefe9ef956240b4577c0fb5cb