Add category to AnalysisCards (#3414)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3414

We've had a few instances come up in meetings last week that I think will make having a category useful:
1. Allow us to group similar analysis together in the UI so everything isn't shown in one tab
2. Allow for easy grouping of analysis in bento notebooks

Additionally, as I was implementing this, there is a TODO on base_client to find a good hueristic for which analysis to show, I think implementing that hueristic will be much easier with category.

I think this is useful enough to warrant the addition because it is distinct from level- which indicates importance - and category should still be ranked by level.

Reviewed By: mpolson64, Cesar-Cardoso

Differential Revision: D69726341