Reorganize tutorials (#3420)
Summary:

Add a new table "Applications and Integrations" to hold our automl tutorial (instead of having it in basic usage).

CristianLara as a side note: Should we improve (or get rid of) the tutorials landing page somehow? Maybe either have it act as some sort of table of contents or just drop people straight into Ax for Ask-tell...

Will leave the decision up to you.

Reviewed By: saitcakmak

Differential Revision: D70252244