Prevent noisy log from MapData
Summary: Tutorial used to be flooded with `/Users/mpolson64/ax/ax/core/map_data.py:216: FutureWarning: The behavior of DataFrame concatenation with empty or all-NA entries is deprecated. In a future version, this will no longer exclude empty or all-NA columns when determining the result dtypes. To retain the old behavior, exclude the relevant entries before the concat operation.`

Differential Revision: D70273329