Expose compute_adhoc method for cross validation plot (#3428)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3428

There was a push last half from folks on the platform team to revamp our analysis code, and it is much crispier than before(yay!). As part of this migration effort we want to exclusively use these new plots moving forward

However, previously these plots did not expose a way to call them directly for adhoc computation. I will be adding this for the plots that need it, and starting with CV.

Reviewed By: sdaulton, mpolson64

Differential Revision: D69754828