Remove status_quo_name input from Adapter
Summary:
Since experiment is always available on the `Adapter`, we can extract it directly from the experiment. For now, the `status_quo_features` input (which is used through input constructors) remains. I'd like to transition that off to the `experiment` as well.

Updated the storage code to remove deprecated kwargs while loading the GNodes & GRs from the db.

Differential Revision: D70292112