Remove status_quo_name input from Adapter
Summary:
Since experiment is always available on the `Adapter` (as of D70103442), we can extract the status quo name directly from the experiment. For now, the `status_quo_features` input (which is used through input constructors) remains. I'd like to transition that off to the `experiment` as well.

Updated the storage code to remove deprecated kwargs while loading the GNodes & GRs from the db.

This is a step in the direction of eliminating redundant inputs within the modeling layer and bringing us closer to relying on the experiment as the source of truth for the inputs that are typically extracted from the experiment.

Differential Revision: D70292112