Allow int-valued noise_std in BenchmarkRunner
Summary:
**Context:**

Yuck! Python typing says an int is acceptable when annotations indicate the need for a float, but `isinstance(3, float)` is False. https://discuss.python.org/t/clarifying-the-float-int-complex-special-case/54018

That was a causing a failure in a function with an `isinstance` check even though there were no complaints from Pyre.

**This PR:**

`isinstance(x, float)` -> `isinstance(x, float | int)`

Reviewed By: Balandat

Differential Revision: D70392395