Allow int-valued noise_std in BenchmarkRunner (#3439)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3439

**Context:**

Yuck! Python typing says an int is acceptable when annotations indicate the need for a float, but `isinstance(3, float)` is False. https://discuss.python.org/t/clarifying-the-float-int-complex-special-case/54018

That was a causing a failure in a function with an `isinstance` check even though there were no complaints from Pyre.

**This PR:**

`isinstance(x, float)` -> `isinstance(x, float | int)`

Reviewed By: Balandat

Differential Revision: D70392395

fbshipit-source-id: c69d6252dbed87e91d42c03c699c386d72daac4e