Simplify Adapter.status_quo_data_by_trial (#3435)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3435

This diff simplifies the implementation of `Adapter.status_quo_data_by_trial`. The helper was prioritizing `status_quo_name`, which always exists if the status quo is set. So, the code block for extracting it based on features was redundant.

It'd be great to extract this directly from the experiment. The main challange is that we currently rely on `Adapter._training_data`, which excludes out of design observations, and that Adapter can have `status_quo_features` that is different than SQ of the experiment (though this is not a critical issue -- typically only trial index differs).

Q: Do we care about excluding out-of-design observations in a world where we expand the model space by default?

Reviewed By: ItsMrLin

Differential Revision: D70336263

fbshipit-source-id: 9836658b816ea46f687e2b26f7695c7db4612b32