Allow for adhoc computation of multiple CVs at one time (#3427)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3427

This is the second ux improvement diff for adhoc cross validation computation. This allows for mulitple metrics to be computed at one time for a single adapter.

It does not currently tile or drop down the metic names. This is a nice improvement we'd like to make, but it wasn't immediately clear how to go about this

Also fixed the CI bars per bernie's catch

Differential Revision: D70200056