`DataLoaderConfig` - Only load last observation of map data by default (#3403)
Summary:

The original purpose of this commit was to ensure that the `Adapter` only loads a single observation by default, even for running and completed map metrics. This makes sure that any method (not-necessarily map-data aware) can be applied by default.

To this end, we need control of the `latest_rows_per_group` of `observations_from_data`. 

In order to avoid increasing the number of arguments to the adapter, we package the new argument and other parameters controlling what data the model is fit to in a new `DataLoaderConfig` dataclass.

Reviewed By: saitcakmak

Differential Revision: D69992533