Minor refactor of insample
Summary:
I'm working on getting an adhoc method exposed for computation of insample_effects on the fly in notebook enviroments. As part of this, I was just getting familiar with this file. This diff simply cleans up unused methods/args, adds comments for clarity and does a minor restructuring of some if/else calls to make clearer.

More substaintial changes to come

Differential Revision: D70445902