Improve ergonomics around early stopping (#3446)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3446

As discussed in Ax/Botorch sync. mark_trial_early_stopped signuture is brought in line with other mark_trial_STATUS methods and BaseTrial now validates that data must be present on a trial when marking it early stopped. This Exception from BaseTrial is a ValueError to stay in line with convention in that module, but is wrapped in an Ax UnsupportedError in Client where we expect exceptions to be Ax classes.

Reviewed By: lena-kashtelyan

Differential Revision: D70397779

fbshipit-source-id: 01f1af2ed7915fcafc7817668a9a6aedb640c522