Improve header for all cards (#3456)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3456

Previously headers used Markdown headings 2 and 3 for title and subtitle respectively, but this was both too big and caused problems in our website build where Markdown headings are automatically interpreted as items to be added to the side bar.

Changed so the title is now just **bold** and the subtitle is normal paragraph text.

Reviewed By: mgarrard

Differential Revision: D70585345