Prevent noisy log from MapData (#3424)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3424

Tutorial used to be flooded with `/Users/mpolson64/ax/ax/core/map_data.py:216: FutureWarning: The behavior of DataFrame concatenation with empty or all-NA entries is deprecated. In a future version, this will no longer exclude empty or all-NA columns when determining the result dtypes. To retain the old behavior, exclude the relevant entries before the concat operation.`

Reviewed By: esantorella

Differential Revision: D70273329

fbshipit-source-id: 0197caa1b90057f6330eb030730936d889484b98