Minor refactor of insample (#3447)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3447

I'm working on getting an adhoc method exposed for computation of insample_effects on the fly in notebook enviroments. As part of this, I was just getting familiar with this file. This diff simply cleans up unused methods/args, adds comments for clarity and does a minor restructuring of some if/else calls to make clearer.

More substaintial changes to come

Reviewed By: Balandat

Differential Revision: D70445902

fbshipit-source-id: 46353e9e3b747079a8c81321dd02a8fe13b0d7c9