Minor fix to MapKeyToFloat (#3463)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3463

Updates config processing to only utilize the specified `parameters` in the `config`. Previously, the default `step` map key was required even if the `config` was specified.

Also updated `__init__` a bit to avoid setting `self.parameters`, since this is overwritten in `super().__init__`. The only changes that propagate are the changes that are made directly to the `config`.

Reviewed By: ltiao

Differential Revision: D70645510