Update pending points to be modified in place for efficiency (#3437)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3437

NOTE: this must be landed at the same time as 4/n

Previously (xxx) we updated the extend pending point method to not modify in place, however, that creates a pretty significant slow down which is unacceptable to waveguide. In order to merge the gen methods, and have all paths go through the new _gen_for_multi_with_multi method (including waveguide) we need to have this method modify in place again.

We include a note in the docstring for this. In follow up diffs we will also:
1. explore making pending points a set instead of a list of observation features
2. remove all external setting of pending points in calls to gen - now we set it inside the gen methods so it's redundant to do so in multiple places
3. try to unify the various pending point utils

Reviewed By: saitcakmak

Differential Revision: D68790218

fbshipit-source-id: e19c9cd8bb35e6825d3769b215e22f5078681c66