Add trial filter option to scatter plot (#3449)
Summary:

This diff is the first in a stack of diffs to both improve the utility of the scatter plot overall and expose an adhoc method for using the scatter plot in a notebook environemnt.

This diff adds the ability to apply a trial filter to scatter data, so taht not all trials will be plotted in a single output. This is particularly important for online experimeation usecases, where trials are typically batch trials.

Forthcoming diffs:
- add adhoc method
- expose label_dict to map metric name ot human readable name
- add ci intervals on the plot
- [backlogged] relativize the plot data
- decide if we like the pareto frontier line

Reviewed By: bernardbeckerman

Differential Revision: D70506659