Extract data from experiment inside Adapter._prepare_observations (#3462)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3462

The `data` (when not explicitly specified) was being extracted from the `experiment` before calling `Adapter._prepare_observations` in `Adapter.__init__` but not in `Adapter._process_and_transform_data`. This was a bug. This diff fixes it by extracting `data` directly inside `Adapter._prepare_observations`, which covers both use cases.

Reviewed By: lena-kashtelyan

Differential Revision: D70503014

fbshipit-source-id: b192aabc0db89e34cea3ad20d7d47aee75cc63ef