Add trial filter option to scatter plot (#3449)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3449

This diff is the first in a stack of diffs to both improve the utility of the scatter plot overall and expose an adhoc method for using the scatter plot in a notebook environemnt.

This diff adds the ability to apply a trial filter to scatter data, so taht not all trials will be plotted in a single output. This is particularly important for online experimeation usecases, where trials are typically batch trials.

Forthcoming diffs:
- add adhoc method
- expose label_dict to map metric name ot human readable name
- add ci intervals on the plot
- [backlogged] relativize the plot data
- decide if we like the pareto frontier line

Reviewed By: bernardbeckerman

Differential Revision: D70506659

fbshipit-source-id: 62de4f93e5c3637248ff03c8a0ec235896d7385d