[Docusaurus] Hide recipes navbar item on unsupported doc versions
The Recipe sections, pages, and sidebar were added to the latest docs but are not present in the existing 0.5.0 version. Unfortunately

Unfortunately the navbar items are not included in Docusaurus's versioning system. This is causing Docusaurus to attempt to show the newly added Recipes navbar item even when viewing the 0.5.0 version (which does not contain recipes) leading to a build failure.

Conditionally adding/removing navbar items is not natively supported by Docusaurus but they do provide a way for us to use our own custom navbar items. Our solution is to create a custom wrapper around the `DocSidebarNavbarItem` that only renders it if the current version is not in a specified list of versions.

Inspired by the official recommendation in this docusaurus github issue: https://github.com/facebook/docusaurus/issues/7227