Add CI to scatter plot (#3476)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3476

This diff adds a 95% CI interval to the scatter plot code. We also do a few other updates: (1) remove color matching for each trial -- i think this is actually not that helpful and we may want to use colors to seperate from modeled vs insample predictions in the future. The hovers take care of the info about trials. (2) adds opacity to CI intervals, and markers for cleaner view (3) cleans up the hover template

Reviewed By: mpolson64

Differential Revision: D70540038

fbshipit-source-id: 1cd6f69f2771bbad2ec322a980bd3890c1f9502c