Move the definition of Scheduler test case into the test file (#3488)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3488

Having test cases specified in a non-test file & target makes it much harder to run & debug specific tests using internal tooling.

The specific structure of this test was perhaps useful when we were considering supporting different subclasses of `GenerationStrategyInterface`, but GSI has since been deprecated and there is no need for this to be defined in a separate file.

Also removed the `_get_generation_strategy_strategy_for_test` method that simply returned back the given GS and cleaned up some of the outdated instructions in the docstring.

bypass-github-export-checks

Reviewed By: mpolson64, paschai

Differential Revision: D70909539

fbshipit-source-id: 2b0b061fb6e52172a18e5eb82eaf27f62d1977d2