Restrict get_target_trial_index to only include trials with status quo (#3485)
Summary:

This helper is typically used to extract the trial index for the status quo features. To ensure that the constructed SQ features corresponds to an actual (trial, arm), the helper is updated to return only the trials that include the SQ arm (and have data for it, as was previously required).

Reviewed By: Balandat

Differential Revision: D70899731