Update Adapter._set_status_quo to extract SQ from experiment with target trial index (#3496)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3496

The goal of this diff is to make `status_quo_features` input of `Adapter` redundant. It is currently constructed by extracting the status quo & the target trial index from the `experiment`, just to pass it into the `Adapter` along with the `experiment`. We can do this directly in the `Adapter`.

To ensure that `Adapter._status_quo` is updated whenever we re-fit the model, also added a call to `_set_status_quo` in `_process_and_transform_data`, which is used when re-fitting the `Adapter` without re-constructing it from scratch.

Reviewed By: lena-kashtelyan

Differential Revision: D70395177

fbshipit-source-id: b0375d3a5fe47c04cc08c3dfaf2ab5b741aa71bd