expose ability to plot adhoc generated points (#3491)
Summary:

This diff exposes a method to call scatter plot adhoc in a bento enviroment. It also exposes the ability to pass in a custom gr and adapter, and see the predictions from that gr in the plot. It does not allow for passing in multiple gr from multiple adapters -- rn you will need to call scatter with each model<>gr combo, I think this is a fair limitation for now.

Forthcoming diffs:
- [will backlog] add ability to provide a single metric and get all metrics plotted against that metric
- [will backlog] expose label_dict to map metric name ot human readable name
- [backlogged] relativize the plot data

Reviewed By: Cesar-Cardoso

Differential Revision: D70580029