Fixes for LaTeX and cell source (#3401)
Summary:
This is a duplicate of the same code change in botorch: https://github.com/pytorch/botorch/pull/2752

# LaTeX in Tutorials

Docusaurus 3 requires math to be wrapped in $ delimiters. We were trying to be backwards compatible with our existing tutorials by wrapping math blocks in $ delimiters while converting to MDX but properly handling every type of math block (which this was failing to do) is beyond the scope of this script.

Since  these tutorials are intended to be consumed by Docusaurus lets just enforce the $ delimiters when writing tutorials.

Before:
<img width="1711" alt="image" src="https://github.com/user-attachments/assets/1d710d83-038b-4f81-907d-d2a93d1eba37" />

After:
<img width="1711" alt="image" src="https://github.com/user-attachments/assets/21f4bc0a-9c52-49b7-88dd-5eef98606050" />

# Tutorial cell source

I was noticing that sometimes (depending on a variety of factors like the editor used to modify the tutorial, whether they had been executed) the conversion script would fail because the cell source would be stored as a list of strings instead of a multi-line string.

From the official nbformat documentation:

> Some fields, such as code input and text output, are characteristically multi-line strings. When these fields are written to disk, they may be written as a list of strings, which should be joined with '' when reading back into memory. In programmatic APIs for working with notebooks (Python, Javascript), these are always re-joined into the original multi-line string. If you intend to work with notebook files directly, you must allow multi-line string fields to be either a string or list of strings.

https://ipython.readthedocs.io/en/3.x/notebook/nbformat.html

Pull Request resolved: https://github.com/facebook/Ax/pull/3401

Reviewed By: mpolson64

Differential Revision: D70990589

Pulled By: CristianLara

fbshipit-source-id: ac189c938bfc24e39dc88a4f5861945b8ddcd6fb