add metric name mapping (#3501)
Summary:

Another nice UX look/feel additiive for the scatter plot is to accept metric mapping to replace long names with human readable names -- this prevents the axes from spilling over.

In the future we should accept and consume this on the UI side as well. But that's a larger discussion related to canonical names

Differential Revision: D70978118