add metric name mapping (#3501)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3501

Another nice UX look/feel additiive for the scatter plot is to accept metric mapping to replace long names with human readable names -- this prevents the axes from spilling over.

In the future we should accept and consume this on the UI side as well. But that's a larger discussion related to canonical names

Reviewed By: mpolson64

Differential Revision: D70978118

fbshipit-source-id: 5c4e2bbb6d0546c0bf0c15e76aecb8caead66269