Clean up derelativize_with_raw_status_quo
Summary:
We were setting `derelativize_with_raw_status_quo` to `True` across the board (in `choose_gs_internal`) because the `Winsorize` transform did not support relative constraints when it was `False`. If a config has only one valid value, it seems simpler to make that the default behavior and eliminate the config. This is exactly what this diff does.

Also cleaned up some long deprecated config options from `Winsorize` and all usage of `derelativize_with_raw_status_quo` across the dispatch utils.

This config was occasionally passed to the `Derelativize` transform as well. When it comes to `Derelativize`, the consensus seems to be that we want to derelativize using the model predictions, so that the model predictions that are compared to the derelativized constraints are on the same scale as the constraint values. So, it doesn't seem to be necessary to use this with `Derelativize` either.

Backwards compatibility: Choose GS does not have any storage associated with it, so no concerns there unless some notebook was explicitly passing this in. The old configs should be loaded and passed into the transforms without issues, as the unexpected kwargs will be silently ignored.

Differential Revision: D71225747