Filter out of design points in best point selection (#3492)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3492

Prereq for solving https://github.com/facebook/Ax/issues/3269

This situation may arise when a user attaches a custom arm (say a baseline or data from a previous experiment) which is intentionally out of design. If a user specifies they are only interested in searching over a specific space we should not return points outside of that space when they request the best trial or pareto frontier.

Note that this should not be an issue for modeled best point selection, where we already are querying the model for best point within the search space (not model space) https://fburl.com/code/e2bw7rnb

Reviewed By: Balandat

Differential Revision: D70920707

fbshipit-source-id: 6c64865a68fd4090bb7a48a9b0b27bb3676e24e2