Allow out of design custom trials on experiment (#3493)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3493

As mentioned in https://github.com/facebook/Ax/issues/3269

We've discovered that attaching an out of design custom arm is a fairly common desire in Ax experiments, typically because a user has existing data from outside of the search space which may be useful for the surrogate model to know about.

This diff makes it so out of design points added via Experiment.attach_trial will continue to be validates, but out of design will now WARN instead of raising an error.

We also add functionality to prevent out of design points from being selected in a previous diff.

Reviewed By: Balandat

Differential Revision: D70920706

fbshipit-source-id: 0e662b3981634cb17478a76987b75e5cd030a35b