Convert Derelativize transform to Adapter._derelativize_optimization_config method
Summary:
`Derelativize` is the only transform that requires access to the `Adapter` (for model predictions) beyond the simple attributes that can be extracted from `experiment`. This change brings us closer to eliminating the dependency of `Transforms` to `Adapter` and replacing it with a simple `experiment` input.

`derelativize_with_raw_status_quo` helper is also deprecated, and replaced with a call to `adapter._derelativize_optimization_config`.

A follow-up diff will remove the `Derelativize` transform, since it simply duplicates what is done in this method.

Differential Revision: D70977939