Forbid acqf options in Ax GeneratorSteps
Summary:
I removed the option by having a conditional inside our botorch model, as I understand it to only impact this model and other models might use this keyword arg to mean something else.

I also believe that the arg was renamed compared to the Task to `acquisition_function_kwargs` and operated under that assumption.

Differential Revision: D71336403