[Docusaurus] Render Recipes nav item iff available in current verison context
In https://github.com/facebook/Ax/pull/3475 we added custom support for rendering navbar items based on the currently selected docs version, but this does not work on non-docs pages. Specifically, the landing page is a "page" not a "doc" (pages are not versioned, only docs) so when on that page we lose context for the currently selected version. Even if we update our logic to render the navbar item when there is no version context, Docusaurus throws an error because it searches through all the possible versions assuming that the targeted sidebar exists in all of them (in our case it does not since the recipes section is newly added)

Docusaurus is also aware of this issue of version context only being available on docs pages, and they use the method `useDocsVersionCandidates()` to determine version candidates that a doc/sidebar a navbar item should point to when the exact selected version is not known. Unfortunately they assume that the doc/sidebar is present in every version so we use that same method to return early if the targeted sidebar is not present in the top version candidate.