Move choose_generation_strategy from ax.preview.modelbridge to ax.preview.api (#3522)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3522

We need to move the new dispatch logic out of ax.preview in preparation for promoting the api module out of ax.preview. There are a number of things to take into consideration while making this move, including:
* Keeping the api module a top level import and not importing it from anywhere else in the Ax codebase
* Continuing to allow choose_generation_strategy to accept a GenerationStrategyConfig rather than all the options and flags individually

After some consideration, the most natural place for this is inside the api.utils module. This is where the other functions that consume Configs and produce Ax objects live.

Reviewed By: saitcakmak

Differential Revision: D70732507

fbshipit-source-id: 2c74d88cc22c29ed3a17435d8fa01dea3d42c4ea