Cleanup map_key_as_features flag (#3532)
Summary:

This flag was being set based on `fit_only_completed_map_metrics`, which was incompatible with the usage of map data in various applications. In addition, it was extracting features from the metadata and adding them to the parameterization, without adding them to the search space. We're cleaning up this option in favor of `MetadataToFloat` transforms and its subclasses such as `MapKeyToFloat`.

Reviewed By: bernardbeckerman

Differential Revision: D71396271