Cleanup map_key_as_features flag (#3532)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3532

This flag was being set based on `fit_only_completed_map_metrics`, which was incompatible with the usage of map data in various applications. In addition, it was extracting features from the metadata and adding them to the parameterization, without adding them to the search space. We're cleaning up this option in favor of `MetadataToFloat` transforms and its subclasses such as `MapKeyToFloat`.

Reviewed By: bernardbeckerman

Differential Revision: D71396271

fbshipit-source-id: 938129033b54e45a56298f264bb04dbf736d4d4b