BilogY transform (#3541)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3541

This adds a bilog transform that is applied automatically to all outcome constraints. The key idea of the bilog transform is to apply a two-sided log-transform that magnifies the constraint boundary, thus making it easier to model that part of the metric space. We modify the transform slightly compared to what is in https://proceedings.mlr.press/v130/eriksson21a/eriksson21a.pdf so that `f(bound) = bound`. This is convenient since it means we don't have to transform the optimization config.

The transform and inverse transforms take the form:
`f(y) = bound + sign(y - bound) log(1 + |y - bound|)`
`f^{-1}(y) = bound + sign(y - bound) (exp((y - bound) sign(y - bound)) - 1)`

The transform is illustrated in the figure below where the constraint bound is `y=5`
{F1976203330}

Reviewed By: Balandat

Differential Revision: D71446538

fbshipit-source-id: 18ce5df0e3a01f42f9f5a7a91a19484d68728654