Support use_posterior_predictive argument in predict methods
Summary: This will allow us to make predictions with observation noise included. We already support this for cross validation. This diff extends the API support to all predict methods on Adapter & Generator classes. Functionally, this is only supported by `BoTorchGenerator` and others will simply warn when posterior predictive is requested.

Differential Revision: D71552963