Support use_posterior_predictive argument in predict methods (#3542)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3542

This will allow us to make predictions with observation noise included. We already support this for cross validation. This diff extends the API support to all predict methods on Adapter & Generator classes. Functionally, this is only supported by `BoTorchGenerator` and others will simply warn when posterior predictive is requested.

Reviewed By: sdaulton

Differential Revision: D71552963

fbshipit-source-id: 5ead5a2143e71a0520c4d15a319467e73e5fb0e5