Warn if raw SQ is out of CI for predictions when derelativizing with model predictions (#3545)
Summary:

As titled. This spawns from some discussion around using raw vs predicted values during derelativization. If the SQ is out of design, we either error out (if no data) or use raw values (if out of design data available). If it is in design, because we use GP models, the model predictions with in-sample points should be pretty close to the raw observations, so we wouldn't expect them to diverge by much. This warning will flag if the two end up diverging unexpectedly.

Differential Revision: D71567208