Update transform storage to utilize strings rather than integers (#3553)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3553

The existing storage for transforms relies on an integer valued registry. The transforms are mapped to the integer values found in the registry, which is JSON encoded along with the transform name, and the decoder looks up the transform class from the reverse registry using the integer.

This comes with a challange when prototyping new transforms. To support them, we may need to add them to the registry with some integer, which may get assigned to some other transform before the transform lands (this is a race condition that shouldn't happen often but it did happen recently). Besides this, if the transform lives in the internal codebase, we need to assign it an integer that doesn't conflict with the OSS registry, then deal with this inconsistency when moving the transform to OSS (since we now want it to have a value that is future & backwards compatible).

But why do we need integers in the first place? This is JSON encoding and the transform name is already being encoded along with the integer. If we updated the reverse registry to map the transform name to the classes, we can just look up the class from the transform name. If we're then moving some transform from internal codebase to OSS, the name is unchanged so we just need to add it to the OSS registry for everything to continue working. No need to deal with integers or race conditions.

Also updated the logic for deprecated & removed transforms to work with the updated logic.

Backwards compatibility: I don't think there are any concerns here. I was happily surprised when I saw that the transform name was already being stored along with the integer value. We will just make use of it when loading old experiments.
Update: Actually, there was a minor backwards compatibility issue, since the previously stored string was `str(transform_type)`, which produces `"<class 'ax.modelbridge.transforms.transform_type'>"` rather than `"transform_type"`. Added a logic to extract `"transform_type"` from there.

Differential Revision: D71590471