Remove sphinx docs for everything *except* API (#3552)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3552

This is something I've discussed with a number of people incl. CristianLara and lena-kashtelyan that I feel would be a positive direction for the usefulness of the website.

I've long suspected that the sphinx docs are not a real value add for our users as they currently stand. They occupy a strange middle ground where they're both more granular than the docs/tutorials on the website, but less legible than reading the code directly on github. They make no effort to distinguish between code that our users should be interacting with and code that exists just for us as helper methods. And, since the overall docstring quality throughout our codebase is mixed the quality of copy on these pages is mixed as well.

By cutting down what gets rendered to just the API we both focus users towards the exact region of the codebase they should be interacting with AND vastly increase the average quality of the docstrings. Developers (as opposed to users) will now need to browse source to find the docstrings for non-API methods that used to be on sphinx *but I suspect they were already doing that anyways*.

Cutting down what gets rendered will increase usability for normal users, likely have zero effect on developers, and will mark our lives marginally easier since we dont have to remember about adding an item to the .rst files

Reviewed By: lena-kashtelyan

Differential Revision: D71578581

fbshipit-source-id: 6d2d826c07e5af35ba167f516867cf1dc86df5d2