GenerationNode: use experiment rather than self.experiment when fitting fallback model
Summary:
We're generating candidates for the `experiment`, so we can use it for fitting the model. It seems somewhat odd to be switching between `self.experiment` and `experiment` for `fit` and `gen` calls. In addition, `self.experiment` creates an unnecessary dependency on the generation strategy, which makes it more difficult to test related functionality.


This brings up the question: A lot of the internal logic around transitions in `GenerationNode` relies on `self.experiment`, which looks it up from the `GenerationStrategy`. Since `GenerationNode` itself now receives the `experiment`, should we just set `GenerationNode.experiment` on each call to `GenerationNode.gen/_fit` rather than accessing to it through `GenerationNode.generation_strategy.experiment`?

Reviewed By: lena-kashtelyan

Differential Revision: D71772493