Add utility to transform per-trial optimization trace to per-step optimization trace (#3560)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3560

Context:

When running benchmarks with MapData, there are multiple ways in which someone might want to analyze the results. By default, `BenchmarkResult.optimization_trace` gets a new entry whenever a trial completes, and it should be compared with respect to `BenchmarkResult.cost_trace`, which tracks cumulative runtime. Another approach is to add a point to the optimization trace each time a step (epoch) completes, and plot against the cumulative number of steps (which would just be range(1, n_steps + 1)).

This diff:

Adds a utility to transform an optimization trace in the former format to an optimization trace in the latter format.

Reviewed By: Balandat

Differential Revision: D71817937

fbshipit-source-id: 927819427d017f3918f8b4c64bc96279820b9c28