Add custom rendered for list[AnalysisCard] (#3568)
Summary:

Since we decided to make Analysis.compute(...) return list[AnalysisCard] we lost the ability to quickly inspect the result of of Analysis.compute(...) or Client.compute_analyses(...). When they returned a card they would simply render the card -- now the render a list of the Card objects string reprs (ex. `[<ax.analysis.plotly.plotly_analysis.PlotlyAnalysisCard at 0x7f4e8fbdfd10>,
 <ax.analysis.markdown.markdown_analysis.MarkdownAnalysisCard at 0x7f4e88bd28a0>,
 <ax.analysis.markdown.markdown_analysis.MarkdownAnalysisCard at 0x7f4e87189040>,
 <ax.analysis.plotly.plotly_analysis.PlotlyAnalysisCard at 0x7f4e87d23e90>,
 <ax.analysis.plotly.plotly_analysis.PlotlyAnalysisCard at 0x7f4e87d463c0>,
 <ax.analysis.analysis.AnalysisCard at 0x7f4e87d46600>]`)

To address this, the first time we create an analysiscard object we register a custom ipython formatter that hijacks the list formatter. It checks (lazily) that every item in the list is an analysiscard and if so renders the html in the same way display_cards does. Note that this does not happen at module import time which is bad practice and can be affected by lazy imports (cinder) magic.

Differential Revision: D71912434